requirements.txt
asgiref==3.5.2
certifi==2022.9.14
Django==4.1
djangorestframework==3.14.0
pip==21.3.1
pytz==2022.4
setuptools==60.2.0
sqlparse==0.4.1
wheel==0.37.1
